104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	unsigned long long int y, sum, num1, num2;
+	unsigned long int y, sum, num1, num2;
 
 	num1 = 0;
 	num2 = 1;
@@ -17,9 +17,9 @@ int main(void)
 		sum = num1 + num2;
 		num1 = num2;
 		if (y < 98)
-			printf("%llu, ", sum);
+			printf("%lu, ", sum);
 		else
-			printf("%llu", sum);
+			printf("%lu", sum);
 		num2 = sum;
 	}
 	printf("\n");
